refactor: PointManager 기능 Point로 이전
