+++ src/main/java/cart/domain/Point.java
@@ -1,5 +1,7 @@
 package cart.domain;
 
+import java.sql.Timestamp;
+import java.time.LocalDateTime;
 import java.util.Objects;
 
 import cart.exception.NegativePointException;
@@ -8,6 +10,8 @@ public class Point {
 
     public static final Point ZERO = Point.valueOf(0);
     private static final int MINIMUM_POINT = 0;
+    private static final int EXPIRED_MONTH = 6;
+    private static final double EARNING_RATE = 0.05;
 
     private final int value;
 
@@ -18,6 +22,21 @@ public class Point {
 
     public static Point valueOf(final int value) {
         return new Point(value);
+    }
+
+    public static Timestamp getExpiredAt(final Timestamp createdAt) {
+        final LocalDateTime localDateTime = createdAt.toLocalDateTime();
+        final LocalDateTime expiredAt = localDateTime.plusMonths(EXPIRED_MONTH);
+        return Timestamp.valueOf(expiredAt);
+    }
+
+    public static double getEarningRate() {
+        return EARNING_RATE * 100;
+    }
+
+    public static Point getEarnedPoint(final Price price) {
+        final int earnedPoint = (int) (EARNING_RATE * price.getValue());
+        return new Point(earnedPoint);
     }
 
     private void validate(final int value) {
