--- src/main/java/cart/domain/PointManager.java
@@ -1,12 +1,0 @@
-package cart.domain;
-
-import java.sql.Timestamp;
-
-public interface PointManager {
-
-    Point getPoint(Price price);
-
-    Timestamp getExpiredAt(Timestamp createdAt);
-
-    double getEarningRate();
-}
